Created webpage that works with local server
Created an index.html that displays the arrow notes in order. Needs a local server at the moment due to the CORS policy as it would be a security risk. One way around this would be to add the json manually in the python script.

Python script now generates a json file with the duration of the note and its direction.